refactor(core): centralize locale all.ts with platform re-exports
Move the aggregated locale map to @videojs/core/i18n/all and keep thin
html/react entry points for existing import paths.
