+++ packages/core/package.json
@@ -39,6 +39,11 @@
       "development": "./dist/dev/i18n.js",
       "default": "./dist/default/i18n.js"
     },
+    "./i18n/all": {
+      "types": "./dist/dev/i18n/all.d.ts",
+      "development": "./dist/dev/i18n/all.js",
+      "default": "./dist/default/i18n/all.js"
+    },
     "./i18n/locales/*": {
       "types": "./dist/dev/i18n/locales/*.d.ts",
       "development": "./dist/dev/i18n/locales/*.js",
